login functions in their own file, currentUser getters and setters in own file
