+++ services/currentUser.js
@@ -1,0 +1,11 @@
+let currentUser;  
+
+function setCurrentUser(user) {  
+  currentUser = user;  
+}  
+
+function getCurrentUser() {  
+  return currentUser;  
+}  
+
+module.exports = { setCurrentUser, getCurrentUser };
